requirements.txt
dnspython~=1.16.0
pymongo~=3.12.0
bson~=0.5.9
werkzeug~=2.0.2
flask~=2.0.2
numpy~=1.24.2
nltk~=3.7
pandas~=1.5.3
gensim~=4.3.1
requests~=2.27.1
bs4~=4.11.2
beautifulsoup4~=4.11.2